Made hyperactive resource a little bit OAuth friendly.
Requires 'oauth' gem for OAuth functionality

Now defers to OAuth to make HTTP requests when an access token is provided. Each HyperactiveResource instance has the attribute access_token which can either be set or not set. If it is not set, the resource behaves as normal. If it is set, special edits are made to the parameters during encoding, and the actual requests are made by the access token/OAuth::Consumer#request.